Implement termination signal handling
Upon receiving a termination signal, sandboxfs should unmount the file
system instead of abruptly exiting and leaving the mount point behind
in an unusable state.  Do so now.

At first I cloned the exact implementation of the Go variant, using the
same message-passing structure between the main program and the thread
handling the signals.  This worked fine but then realized (thanks to a
"note" in the original code!) that we can do better in Rust because of
the ability to control the signal mask via the nix crate.

Note, however, that the way signals are handled may still seem odd in
cases (like e.g. ignoring all signals after the first one is received).
This is intentional to keep the behavior in sync with the Go code and
with what is currently tested.