Make ShareableFile actually unblock reads (#51)
As it turns out, closing the file descriptor on which a read syscall is
blocked doesn't actually cause the read to unblock.  According to some
documentation I found online for Linux, this is because the read blocks
on a the file itself so closing the descriptor has no effect on the read.
I suspect a similar reason applies to macOS, where I'm seeing this now.

To fix this, use select instead before attempting the read to ensure that
the read will complete.  But the same applies to select: closing the file
handle on which select waits is undefined behavior and, while it unblocks
the select on macOS, it doesn't on Linux.  So add a pipe to communicate
between the threads, and use a write to the pipe to unblock the select
and let other threads know that the file is closing.

I have no clue how I concluded that this worked in the past.  Obviously
the test was very wrong (writing to the FIFO actually made the read
unblock regardless of what we did elsewhere)... but I had tested this on
the real sandboxfs code after implementing reconfigurations and things
seemed fine.  I was probably doing the wrong thing too, and now that
singal handling is in place, the problem becomes more obvious.