Add parsing and processing of reconfigurations
This introduces a new "reconfig" module that implements the parsing of
reconfiguration requests in their serialized JSON format and adds a
ReconfigurableFS trait that abstracts away the operations that need to
be supported to apply these requests.

The reconfiguration format is suboptimal in many ways but tracks
exactly what the Go variant of this code does so that we can remain
compatible with it for the time being.  Ideally we should probably use
some form of true RPC system, like gRPC or HTTP handlers, but if we
do that we need to be careful about performance.

The code added here is still unused.