Implement reconfiguration support
This hooks up the processing of reconfigurations from the reconfig.rs
module into a separate thread that listens for requests on the file
provided in --input and responds to them via the file provided in
--output.

Note that reconfigurations are "broken" when compared to the Go variant:
the FUSE crate we use does _not_ currently support kernel cache
invalidations... so unmap operations do not work.  I have implemented
them as much as possible but they do not truly work.  I'm willing to
live with this limitation for now though: the primary use case we
target, Bazel, does not require unmaps to work because each sandboxed
action runs within a fresh and unique subtree of the sandboxfs instance.

It's possible that we can make unmaps work by disabling all kernel
caches, at least for testing purposes, but we do not currently have the
ability of passing arbitrary FUSE options at mount time... so will
investigate later to not delay this any further.