Implement renames and moves
Handling renames within a single directory is easy, but supporting a move
across directories is difficult because we need to lock two nodes and it
is good to not leak the implications of this through the Node trait.

To support the latter case, I have split the rename operation in two: one
for the responsibilities of the source node and one for the target node.
This approach allows keeping the Node interface and the Dir type
oblivious to the various Node types that exist (i.e. we don't need
introspection as the Go variant of this code needed).  The downside is
that the need for the source/target method split looks strange, and in
the source side, we have to support "rolling back" a failed move.
Having seen the two possible implementations, I think this is clearer.

This design is also known to deadlock under concurrent renames... but the
FUSE library we use serializes all requests so we can never hit that
condition.