Ensure we don't reuse inodes after deletion
Make sure to remove entries from the node cache (underlying path to
node) when an underlying path is deleted.  This prevents reusing the
data of a deleted file when it is recreated in the same place.

This fixes the TestReadWrite_InodeReassignedAfterRecreation test and
extends it to cover the case of directories, which we know is different
because we currently cannot cache directories.