Delegate permissions checks to the kernel
Pass "-o default_permissions" to the FUSE mount operation to let the
kernel handle permissions checks.  Based on previous experience with
the Go code, this can yield a significant performance boost.

But also, this fixes the TestReadOnly_Access test on Linux, which was
not working properly at all because we don't implement access checks
on our own.