Unmount the file system by calling unmount(8)
When we receive a signal, we have to unmount the file system in order to
unblock the main fuse loop.  We were previously doing this by calling the
(private) `fuse::unmount` function... but this wasn't working properly on
Linux: this function does not expose any errors that happen during the
call to fusermount and thus our unmount retries weren't working correctly
on a busy file system.

To fix this, simply delegate the unmounting operation to the userspace
tool (either unmount(8) or fusermount(8)).  Trying to do this in-process
is very difficult due to portability differences, and this needn't be
efficient because this code only exists to deal with a corner case.

Fixes TestSignal_QueuedWhileInUse on Linux.