Replace the Go installer script with a Makefile
Now that the Rust variant of sandboxfs will take over this package,
let's try to minimize the use of Go to simplify the installation
process.  Having to have Bazel and Go in place just to run a trivial
installer script is overkill, so replace it with a more traditional
"make install" approach.  (This actually restores much of the
configure/Makefile code we had before the Bazel transition.)

Note that, contrary to what we did with Go before, this does not
intend to replace the standard toolchain: the Makefile still delegates
to Cargo for the build as building things on our own is too complex,
not worth it, and not integrating with the Cargo ecosystem would be a
mistake.