Stop using Bazel to run the tests
Now that we don't care any longer about building the sandboxfs binary with
Go, we don't have to worry about reproducibility as much as before (which
was the main selling point for using Bazel with Go).  We now only use Go to
run the integration tests and linting, so we can be a bit more lax.

This change switches to using the standard Go tools to run the tests and
removes a lot of complexity we had to deal with VSCode.  The editor now
"just works" again as we don't rely on the unmatained bazel-gopath tool.