Stop using Bazel to lint sources
The linter tool is pretty simple, so we can run it directly from the
Makefile.  This is much easier... and makes interactive lint executions
as part of the Git pre-commit hook tolerable again: running Bazel there
was incredibly painful and it became worse recently because some tool
(I think it was Gazelle) started pulling in protobuf, which takes ages
to build.