Remove support to test the Go variant of sandboxfs
This drops all hacks that were put in place within the integration tests
to cope with the Go/Rust duality.

I'm keeping some of those hacks in reconfiguration_test.go though, as they
exist to prevent testing the unmappings that are not known to work.  That
code is worth keeping as is because we do want unmappings to work again.