Allow disabling Go usage and disable it for install
Fix the behavior of the --goroot flag so that it recognizes a "none"
value that cleanly disables Go usage from the build.

To ensure this works, disable Go usage during the install Travis
configuration: we don't need to build the tests nor lint the source
tree in this case, so we don't need Go.