Verify that sandboxfs works well for Bazel builds
Our primary customer is Bazel so let's make sure sandboxfs works well
for Bazel builds by adding a Travis configuration that builds Bazel
with itself and with sandboxfs support enabled.

This is intended to catch real-life problems in sandboxfs: our unit
and integration tests are extensive, but they cannot guarantee that a
build, with its complexity in parallelism and different access patterns,
will work.

Note that this Travis job only runs from cron-triggered builds because
of their high cost.