Fix readdir of long directories
It turns out that ReplyDirectory::add() returns a boolean that indicates
whether the buffer for the reply to the kernel is full or not -- and we
were ignoring this completely.  As a result, readdirs of long directories
were truncated, leading to very strange build failures within Bazel (e.g.
because some jar files lacked entries late in the build).

Fix this by precomputing the full reply to the kernel on the first
readdir call, caching the output, and then returning it to the kernel in
smaller chunks.