Return correct flags on open/create replies
It is not good to reply to an open/create request with the same input flags
that were provided.  Doing so confuses the kernel... which causes mmap to
malfunction in odd ways.  Simply return 0, which is what the Go version
was doing.

This also adds an mmap test, which is not only tied to this change, but
which this change fixes.  This test tries to replicate what "ranlib" does
on macOS, which was very unhappy with our previous behavior and crashed
immediately.  Took me over a day to figure out what was going on...