Support optional Rust features in build infrastructure
In order to support introducing an optional dependency on gperftools,
which we'll do by means of a Rust feature, we need to teach our build
machinery about such optionality.

To make this happen, add a new --features flag to the configure script
so that Cargo invocations and integration tests get a consistent view of
this value, and then teach both the integration tests and the Travis
scripts about this.