Revert to using fuse-0.3
We cannot publish sandboxfs as a crate if it depends on non-published
crates, so we must pin fuse to 0.3 and avoid code that relies on the
Eq comparator for FileAttr and FileType that I contributed upstream
and that hasn't made it into any release yet.

This reverts 67dd2d74c510c49ea35b123cbcde31bce001041d.