Add prebuilt binaries for Linux
This adds a script to generate a tgz file with a prebuilt installation
for Linux that should mostly work without a lot of trouble on various
systems, and documents it in our installation instructions.

As part of the documentation changes, improve the existing content for
other areas to be more specific about which packages are needed when.