Turn reconfiguration protocol into JSON streams
For reconfiguration requests, remove the line-based protocol that expected
one request per line followed by a blank line and replace it with a stream
of JSON messages.

For reconfiguration resposes, remove the text-based answering mechanism
and replace it with a stream of JSON messages as well.

As a side-effect of this change, we cannot really tolerate invalid input
any longer as there is no way to recover the stream once that happens.
Therefore, any invalid input results in a fatal error that causes sandboxfs
to stop accepting reconfigurations.