Tag map and unmap reconfigs with a root directory
This turns each map and unmap reconfiguration step into a list of
mappings, all tagged with a single "root" directory. All mappings in
the same step start in the same directory.

This will allow us to make reconfiguration operations much more efficient
by skipping the lookup of the root directory for each entry to apply.