Use snake case for JSON mappings
The reason JSON mappings used CamelCase in their field names is historical
and dates back to when the protocol was first introduced in Go.

Given that I've already changed the protocol for various other reasons, it
doesn't hurt to also fix this legacy issue and use snake case plus the same
internal names we use now for each mapping part.

I'm keeping Map and Unmap in CamelCase because they represent something
else (an "RPC"), which feels like a different thing than a field.