Simplify reconfigurations to deal with sandboxes
Stop the recent madness that went into the reconfiguration protocol and
vastly simplify it by reducing its power.  Instead of dealing with paths
to map and unmap, deal with "sandboxes", which are just top-level
directories within the mount point that contain mappings.

This strips power off from sandboxfs, but nothing of value is lost: I
cannot imagine a case where arbitrary reconfigurations are useful, and
this makes the whole thing simpler.