Process reconfigurations in parallel
Reconfiguration operations are, unfortunately, not in-memory only at this
point given that they require stating the file system mappings in order
to create the nodes.

Because of this, processing reconfiguration operations is not quick, and
if we process them serially, we penalize the parallelism that the caller
might have (primarily Bazel).

To address this, use a configurable number of threads to apply the
reconfigurations, and process each request in a separate thread.

In the future, we should try to avoiding any I/O during reconfiguration
operations.  To do this, we need to make node creation only require a
type, not the whole file metadata, which in turn requires making the
getattr operation lazy and extending the protocol to carry the node types
(which Bazel might already know).