Update a node's size during writes
When writing to a file descriptor that had been dupped and then closed,
the writes would be "lost" because we did not update the file's size.
This only happened under this condition because, when a file is deleted,
we start serving getattr requests purely from memory (which then get a
stale size value); and because when a file is dupped, the kernel behaves
differently.  (I suspect the duplicated handle gets a copy of the bad
size or something like that but I haven't investigated; sandboxfs's
behavior was obviously buggy.)

This fixes bmake -jN builds when pointing TMPDIR into sandboxfs.  These
builds would seem to make progress but none of the subprocesses executed
at all.  The strange thing is that bmake reported no errors, and it seems
the syscalls themselves didn't either.