Add support for extended attributes
These turn out to be necessary when running tar inside sandboxfs on
macOS.  tar uses the copyfile() call and asks it to copy extended
attributes, and returns an error if it cannot do that.  Only seems
to happen when running as root, however, which is a bit strange.

I'm not making support for xattrs conditional yet, but I suspect
a blanket implementation like this may be detrimental to performance
for the Bazel use case.  If that turns out to be true when I am able
to measure it, I'll add a flag.