Use prefix encoding for paths
Minimize reconfiguration requests by encoding paths as a pair of a prefix
identifier and a basename.  The prefix identifier is simply a number that
points to a path, so that the same number can be reused for all paths in
the same directory.

Prefixes are valid for the entire duration of a sandboxfs session so that
different configuration requests can refer to previously-defined prefixes.
Each request can then introduce new prefixes.