Make xattrs support optional
Add a new --xattrs option to explicitly enable support for extended
attributes, and make it disabled by default.

Support for these has been problematic so far, is still incomplete due
to defficiencies in the APIs we use, and I suspect can have a negative
performance impact on Bazel builds (because the kernel issues a lot of
xattr requests to the daemon that can otherwise be skipped).