Disable the node cache and add a flag to enable it
The node cache has turned out to be problematic, especially on macOS where
we get Java crashes when the toolchain is mapped under multiple sandboxes
and the first sandbox in which it was mapped disappears.

I should probably just remove the cache implementation altogether, but I'm
still nervous about losing what-I-think-is a pretty important performance
optimization.  Furthermore, 0.2.0 has already piled up too many changes,
so doing this blindly would be risky.  Therefore, I'm adding a --node_cache
flag to bring the cache back, but this is probably going to go away to
significantly simplify the code.