Test against bazel 3.0.0rc2
This is the first release with proper support for the sandboxfs 0.2.0
protocol, so start using it to validate our integration with Bazel.

While doing this, bump the timeouts for the install phase, as we now
need to download some larger artifacts, and make sure that the install
phase downloads everything we need (without postponing downloads to
the build phase).  Unfortunately, I've had to disable the osx build
because I was unable to fit it within Travis' timeout.