Refactor how we handle inode operations
Instead of intermixing calls to the nodes and error handling inside the
implementation of the vnops, add helper functions that return Result
objects and make the vnops simply deal with the kernel replies.

Other than making the code simpler to follow, I'm doing this because
find_node will need to change to support returning errors (once I fix
unmap to actually erase nodes).  Once that happens, having to add extra
error logic everywhere will be brittle and make things even harder to
follow.

And to prove this, I've simplified all the logic that called find_node
and then verified if the nodes were writable behind a find_writable_node
call that returns a Result.