Make unmap actually delete inodes
We must do this to keep sandboxfs memory usage under control.

During a large ObjC build from Bazel, I noticed sandboxfs growing to
50GB because that build creates thousands of very large sandboxes.
Even if the build unmaps each of them as it's done with each action,
sandboxfs was keeping track of all inodes.  With this change, memory
consumption drops to a few hundred MBs during the same build.