Adopt go modules
This allows us to remove some logic in the configure script to download
unversioned packages and to fix some problems I was observing with the
VSCode integration.

As part of this, pin bazil.org/fuse to a version that still works on
macOS and go 1.12.  I'll be removing this soon anyway.